Return promises when loading MontageFrame owner
This adds support for creating instances of the owner component as
specified by the serialization given to the MontageFrame. As this
process may be asynchronous the MontageFrame will return a promise for
the owner component created when load is called.

Clients should assume the montageFrame is not ready until the owner
promise is fulfilled.

I have also added some specs for the MontageFrame to exercise the API.
There was some fairly disruptive shuffling of responsibilities as the
MontageFrame and the internal Frame object come into their own.