Provide addObject API to MontageFrame
This works much like (and in fact now does the heavy lifting of)
the addComponent API. This is how you would add anything that is not
a component to the owner component.

Still need to work out how it ends up in the template, how properties
are configured, and so on.

Rudimentary specs have been added.