Replace legacy frame with shell application
This completes the replacement of the old frame object with the new
shell application for hosting components.

The authoring document has also been replaced with a project if only to
reduce the number of "documents" we have to juggle.

Some APIs have probably gone missing from the montageFrame and I expect
more will as I consider promoting the component controller to a place of
prominence.