Improve mouse-driven selection behavior
Don't change selection for any old event, only left clicks will select.

Also added stubs for places to handle drilling down and selection
expanding and contracting.

This is a step towards redefining selectedObjects to be "the objects
that are currently selected" instead of "the list of objects that
have been selected"

The recent selections concept should be a different collection
altogether.

Of course, I have no idea how we'll handle multiple selection in a
single inspector.