Move editing APIs to EditingDocument
This is what is the workbench/editingFrame promise to fulfill when
instructed to load a reel. This simplifies the loading process such
that no editing APIs are available until the editingDocument is
initialized.

The EditingDocument still relies on the EditingController née
ComponentController to do its work if for no other reason than that
body of functionality has moved around a bit already and it's nice to
have it all encapsulated as the final home of the editing APIs
settles down.

The EditingDocument also is now the result of asking to load, instead
of the thing you setup to load, which is much cleaner conceptually.

Now it's a reelUrl + packageUrl all the way through the APIs until
we actually load them onto the stage.

Tests need to be backfilled to lock in functionality from this point.