Retrieve original serialization to populate model
The editing document now creates the template from the serialization
of the ownerComponent.

While we already have a reference to the ownerComponent, fetching the
template is asynchronous and I didn't want an async init; the burden
of doing the asynchronous work is placed upon the caller of the
initialization instead. I might revisit this once we have a better
handle on the async patterns we want to promote.