Introduce EditingProxy Model
At this point objects on the stage that are present within
the owner's template are represented by EditingProxy objects.

The EditingDocument houses the Editing Proxies and assorted editing
APIs. Many of these are still housed in the EditingController for
organization purposes more or less at this point; seeing as the
EditingDocument is the lowest-level editing API object yielded by
palette to any editor relying, the EditingDocument should be
considered the expert on editing.

When making selections on the stage you are given an editing proxy.

Consumers of what were objects live on the stage need to adjust to
accepting the proxies instead. In cases where the actual live object
is needed the proxy does provide a reference.

No edits should be made directly on the editing proxies themselves,
instead the edits should be performed by the EditingDocument, for
now at least.

The EditingDocument will eventually ignore the live stage when
saving, instead relying on the EditingProxies to power the
serialization.