Add removeComponent as inverse to addComponent
They're undoable, but UndoManager will need to accept living in the
brave new world of promises. It can't expect an undo to trigger another
undo addition during the same turn, resulting in a "redo" stack
addition.

Since the original undo process finished, the next undo is treated as
an undo itself. Need to add a way to accommodate async operations
being undoable/redoable of course.