Create base editing-document
Further downplay the importance of a reelUrl/component as a thing worth
considering instead opting to promote the less opinionated fileUrl.

Palette continues to provide the building blocks for editing reels with
a somewhat specialized reelEditing components and a reel-document, while
providing some leeway for others to specialize in other directions.