Add option to select objects on addition
I'm not sure where to draw the line on this for things such as adding
bindings. Currently, I think clients of palette can set the
selectObjectsOnAddition to true or false, and momentarily override it
while performing an addition they don't want to follow the default.

Technically this exact same behavior could be handled at the editor
level, I'm putting it here as I think it would be convenient.