Make editing documents load their own model data
In the case of the ReelDocument we load the reel's template and populate
the authoritative editing model independent of any actual live usage of
the template.

The ReelDocument does allow for eventual association with a live
instance in order to reflect changes to the editing model.

This significantly changes the editing UI componentry which was
previously responsible for building the editing document from the live
template, now it merely needs to return a reference to the live objects
of interest.

The most significant improvement with this approach is that we can load
an editing document and actually edit it independent of having a view
of the changes. This is made possible by the realization that we don't
ever want to use the live objects or live rendered DOM to drive any
part of the final result from the editingDocument.

The EditingController has not yet taken advantage of this new approach;
it still performs edits exactly the someway as before.
edits on the live stage.