Add concept of editor component to document
In short, the preferred component type for presenting and the actual
component currently presenting this document.

I'm sticking with the idea that a single component is coordinating
the presentation of this document; if it enlists other components
to aid in doing so that's its own business.

Additionally, a single document is presented by a single component at
any given time. This seems like a sane expectation for the time being.