Stop rebuilding template after every single edit
This was necessary when we were using the template as a model, but now
it's really the end result that needs to be saved from the editingModel;
there's no need to update it so frequently.

This should improve performance a bit while making frequent edits.

We'll need to promote (and rename) this method as we now expect
interested parties to update the template on demand as they see fit.