Move all properties dispatching to EditingDocument
All dispatching related to changes in properties were in the
EditingProxy, to mimic the same model we have on filament, where
everything is initiated by the ReelDocument we're moving all these
events to the EditingDocument.
One of the main reasons for this change is related to the undo abilities
that are managed by the EditingDocument.