Change Document loading API to init(...).load()
The previous API was a constructor load() method that would fetch
async data from a data reader and would call init()
with the data derived from the reader.

This new API tries to simplify the loading / initialization API by
initializing all the variables with init(...) and preform all async
operations on the load().

There is also a change to use a DataSource object instead of the
dataReader / dataWriter combo that was previously given to load / save.
This new object is so important to the functioning of the Document that
it needs to be given at init time.

Using the DataSource to write the contents on save is not yet
implemented, but it will be in the near future.