Remove unnecessary needsDraw
This was actually the root cause of a fairly significant and annoying
bug, FIL-643; properties for newly added objects would be set to
the cast value of the specific property inspector being used.

This was unexpected.

Doesn't appear that we need to update the tooltips to update. More
advanced inspectors may need to schedule a draw when their value
changes.